+++ lib/esm/index.d.ts
@@ -1,5 +1,0 @@
-import toOrdinal from './toOrdinal';
-import getOrdinal from './getOrdinal';
-import buildBetParams from './buildBetParams';
-import getTeamsAndParams from './getTeamsAndParams';
-export { toOrdinal, getOrdinal, buildBetParams, getTeamsAndParams, };