+++ src/translationByLocale.ts
@@ -1,0 +1,6 @@
+export default function translationByLocale<T>(
+    translations: { [lang: string]: T },
+    loc: string,
+  ): T | undefined {
+    return translations[loc] || translations[loc.split('-')[0]] || translations.en;
+  }